Error fix

+++ src/cosim/slave_simulator.cpp
@@ -241,7 +241,7 @@ public:
             references_filtered_.clear();
             values_filtered_.clear();
 
-            for (int i = 0; i < references_.size(); i++) {
+            for (size_t i = 0; i < references_.size(); i++) {
                 auto& ref = references_.at(i);
                 if ((*filter)(ref)) {
                     references_filtered_.push_back(ref);
